Added contributors_json.js && updated old script to contributors_md.js
